Title: 올바른 등식 만들기, Time: 135ms, Memory: 10MB, Status: Passed - Codetree

+++ 260319/올바른 등식 만들기/right-equality.java
@@ -11,7 +11,7 @@ public class Main {
             num[i] = sc.nextInt();
         }
         
-        int[][] dp=new int[N+1][41];
+        long[][] dp=new long[N+1][41];
         // for(int i=1;i<=N;i++) Arrays.fill(dp[i], -1);
         
         dp[1][num[1]+OFFSET]+=1;
